Add Creature AI
Add attackCheck() that decides if an enemy Creature is close enough to attack. Add member variable attackRange that holds the distance that a Creature can attack from. The attackRange is the same for the X, Y and Z directions. Change dealDamage() to be private so that the only way to deal damage is to call attackCheck(). This prevents a Creature from dealing damage to another Creature when it is out of attacking range.