Add list of landscape and draw them
Add list of type Entity to represent landscape. Import fbx file as placeholder for future landscape model. Add foreach loop in Draw() to draw a landscape object for each landscape. Add foreach loops to call update() on landscape so that they will update their physics. Make velocity vector public so that it can be modified on collision. Add collision check with the ground for enemy and landscape.